+++ src/mixins/general/course-structure.ts
@@ -13,64 +13,6 @@ import {
 import { chapterSlugRoutingIssueAvoid } from '@/mixins/general/course-chapters'
 
 /**
- * @description breaks nextStep property into array of numbers,
- *   used for converting old navigation into new
- * @param block course content block containing nextStep property
- */
-export const legacyContentStepsTransform = (block: LegacyContentBlock): number[] | number => {
-  if (block.nextStep === null) {
-    return null
-  }
-  const stepsArray = block.nextStep.split(',')
-  const stepsInt = stepsArray.map((step) => parseInt(step) - 1)
-  return stepsInt.length > 1 ? stepsInt : stepsInt[0] // return array if more than one step
-}
-
-/**
- * @description transforms follow property from number to id
- * @param courseChapters array of course chapters, result of setCourseContentAndNav mutation
- */
-export const legacyContentFollowTransform = (courseChapters: CourseNavigationItemBlock[]): void => {
-  courseChapters.forEach(chapter => {
-    const transformNumberToId = (follow: number): string => {
-      return courseChapters[follow] ? courseChapters[follow].id : null
-    }
-    if (Array.isArray(chapter.follow)) {
-      chapter.follow.forEach((follow: any, i: number) => {
-        chapter.follow[i] = transformNumberToId(follow)
-      })
-    } else {
-      // @ts-ignore
-      chapter.follow = transformNumberToId(chapter.follow)
-    }
-  })
-}
-
-/**
- * @description descent into course navigation structure and add routes to array
- * @param structure - course navigation structure
- * @param currentPath - current path in course navigation structure
- * @param routes - array of routes, will be modified
- */
-const courseRoutesCollect = (
-  structure: any,
-  currentPath: string,
-  routes: Array<[route: string, id: string]>
-) => {
-  const currentPathSnippet = currentPath ? currentPath + '/' : ''
-  if (structure instanceof Array) { // structure is CourseNavigationItem[]
-    structure.forEach((item) => {
-      courseRoutesCollect(item, currentPath, routes)
-      // routes.push([currentPath + '/' + item.slug ? item.slug : item.id, item.id]) // can have alternative path
-    })
-  } else if (structure.isChapter) { // structure is CourseNavigationItemChapter
-    courseRoutesCollect(structure.children, currentPathSnippet + structure.slug, routes)
-  } else if (!routes.some(el => el[1] === structure.id)) { // add path to routes if id is not already there
-    routes.push([currentPathSnippet + structure.slug, structure.id])
-  }
-}
-
-/**
  * @description traverses chapters, extracts object that has slugs as keys and chapterNames as values
  * @param chapters courseChapters to inspect
  */
@@ -87,6 +29,21 @@ export const courseChaptersCollect = (chapters: CourseNavigationItem[]): object 
     return chapterNames
   })
   return chapterNames
+}
+
+/**
+ * @description transform a content block into a course chapter
+ * @param block to be transformed into chapter
+ */
+export const courseContentBlockToNavItemTransform = (block: ContentBlock): CourseNavigationItemBlock => {
+  const slug = chapterSlugRoutingIssueAvoid(block.title.text, block.name)
+  return {
+    id: block.id,
+    isChapter: false,
+    slug,
+    type: block.name,
+    follow: null
+  }
 }
 
 /**
@@ -117,38 +74,6 @@ export const courseContentIdGet = (courseNav: any, mode: string): string => {
 }
 
 /**
- * @description transform a content block into a course chapter
- * @param block to be transformed into chapter
- */
-export const courseContentBlockToNavItemTransform = (block: ContentBlock): CourseNavigationItemBlock => {
-  const slug = chapterSlugRoutingIssueAvoid(block.title.text, block.name)
-  return {
-    id: block.id,
-    isChapter: false,
-    slug,
-    type: block.name,
-    follow: null
-  }
-}
-
-/**
- * @description traverse course nav object, return set of all paths
- * @param courseNav course navigation object
- * @returns list of tuples with route and id
- */
-export const coursePathsGet =
-  (courseNav: CourseNavigationItem[]):
-    any[] => {
-    const routes = [] // instantiate here because of recursion
-    const start = courseContentIdGet(courseNav, 'first')
-    if (start) {
-      routes.push(['', start])
-    }
-    courseRoutesCollect(courseNav, '', routes)
-    return routes
-  }
-
-/**
  * @description traverse course structure object, extracting all content ids and
  *  writing them to `ids` object
  * @param structure course structure object
@@ -162,40 +87,6 @@ export const courseContentIdsExtract = (structure: any, ids: any): void => {
     structure.children.forEach((item: CourseNavigationItem) => courseContentIdsExtract(item, ids)) // recursively call for each child
   } else {
     ids[structure.id] = {}
-  }
-}
-
-/**
- * @description find chapters with only one item and remove slug
- * @author cmc
- * @since v1.3.0
- * @param routes list of routes
- * @param subChapterSlug slug to keep (chapter name or block title)
- */
-const courseSlugsTrim = (
-  routes: Array<[route: string, id: string]>,
-  subChapterSlug: string
-): void => {
-  if (subChapterSlug === 'chapter' || subChapterSlug === 'block') {
-    routes.forEach((route: [string, string]) => {
-      const pathList = route[0].split('/') // split path into list
-      if (pathList.length >= 2) { // more than one member => subchapters
-        const subChapter = routes.filter(
-          e => e[0].includes(pathList.slice(0, -1).join('/'))
-        )// find all subchapters i.e. all paths with same prefix
-        // @ts-ignore
-        if (subChapter.length === 1) { // only one subchapter => remove slug from path
-          const pathList = route[0].split('/')
-          if (subChapterSlug === 'chapter') { // keep chapter slug
-            route[0] = pathList.slice(0, -1).join('/')
-          } else { // keep block slug
-            route[0] = pathList.length > 2
-              ? pathList.slice(0, -2).join('/') + '/' + pathList.slice(-1)
-              : '' + pathList.slice(-1)
-          }
-        }
-      }
-    })
   }
 }
 
@@ -221,4 +112,76 @@ export const courseDestructure = (
     courseSlugsTrim(routes, subChapterSlug)
   }
   return [ids, routes]
+}
+
+/**
+ * @description traverse course nav object, return set of all paths
+ * @param courseNav course navigation object
+ * @returns list of tuples with route and id
+ */
+export const coursePathsGet = (courseNav: CourseNavigationItem[]): { [id: string]: string[] } => {
+  const routes = {} // instantiate here because of recursion
+  // const start = courseContentIdGet(courseNav, 'first')
+  // if (start) {
+  //   routes.push(['', start])
+  // }
+  courseRoutesCollect(courseNav, '', routes)
+  // return routes
+  return {}
+}
+
+/**
+ * @description descent into course navigation structure and add routes to array
+ * @param structure - course navigation structure
+ * @param currentPath - current path in course navigation structure
+ * @param routes - array of routes, will be modified
+ */
+const courseRoutesCollect = (
+  structure: any,
+  currentPath: string[],
+  routes: object
+) => {
+  if (structure instanceof Array) { // structure is CourseNavigationItem[]
+    structure.forEach((item) => {
+      courseRoutesCollect(item, currentPath, routes)
+      // routes.push([currentPath + '/' + item.slug ? item.slug : item.id, item.id]) // can have alternative path
+    })
+  } else if (structure.isChapter) { // structure is CourseNavigationItemChapter
+    courseRoutesCollect(structure.children, currentPath.push(structure.slug), routes)
+  } else if (!Object.keys(routes).some(el => el === structure.id)) { // add path to routes if id is not already there
+    routes[structure.id] = currentPath.push(structure.slug)
+  }
+}
+
+/**
+ * @description find chapters with only one item and remove slug
+ * @author cmc
+ * @since v1.3.0
+ * @param routes list of routes
+ * @param subChapterSlug slug to keep (chapter name or block title)
+ */
+const courseSlugsTrim = (
+  routes: { [id: string]: string[] },
+  subChapterSlug: string
+): void => {
+  if (subChapterSlug === 'chapter' || subChapterSlug === 'block') {
+    Object.keys(routes).forEach((id: string, i) => {
+      if (routes[id].length >= 2) { // more than one member => subchapters
+        const subChapter = routes.filter(
+          e => e[0].includes(pathList.slice(0, -1).join('/'))
+        )// find all subchapters i.e. all paths with same prefix
+        // @ts-ignore
+        if (subChapter.length === 1) { // only one subchapter => remove slug from path
+          const pathList = route[0].split('/')
+          if (subChapterSlug === 'chapter') { // keep chapter slug
+            route[0] = pathList.slice(0, -1).join('/')
+          } else { // keep block slug
+            route[0] = pathList.length > 2
+              ? pathList.slice(0, -2).join('/') + '/' + pathList.slice(-1)
+              : '' + pathList.slice(-1)
+          }
+        }
+      }
+    })
+  }
 }
