+++ src/mixins/general/legacy-content.ts
@@ -1,0 +1,41 @@
+/**
+ * @file contains transformations from old course structure into new one
+ * @author cmc
+ * @since v1.3.0
+ */
+
+import { CourseNavigationItemBlock, LegacyContentBlock } from '@/mixins/types/course-structure'
+
+/**
+ * @description breaks nextStep property into array of numbers,
+ *   used for converting old navigation into new
+ * @param block course content block containing nextStep property
+ */
+export const legacyContentStepsTransform = (block: LegacyContentBlock): number[] | number => {
+  if (block.nextStep === null) {
+    return null
+  }
+  const stepsArray = block.nextStep.split(',')
+  const stepsInt = stepsArray.map((step) => parseInt(step) - 1)
+  return stepsInt.length > 1 ? stepsInt : stepsInt[0] // return array if more than one step
+}
+
+/**
+ * @description transforms follow property from number to id
+ * @param courseChapters array of course chapters, result of setCourseContentAndNav mutation
+ */
+export const legacyContentFollowTransform = (courseChapters: CourseNavigationItemBlock[]): void => {
+  courseChapters.forEach(chapter => {
+    const transformNumberToId = (follow: number): string => {
+      return courseChapters[follow] ? courseChapters[follow].id : null
+    }
+    if (Array.isArray(chapter.follow)) {
+      chapter.follow.forEach((follow: any, i: number) => {
+        chapter.follow[i] = transformNumberToId(follow)
+      })
+    } else {
+      // @ts-ignore
+      chapter.follow = transformNumberToId(chapter.follow)
+    }
+  })
+}
