+++ src/mixins/vue/plugin-view.ts
@@ -5,10 +5,12 @@
  * Since: v1.1.0
  */
 
+import { mapGetters } from 'vuex'
+
 export default {
   props: {
-    onFinish: {
-      type: [Function, Array],
+    nextContent: {
+      type: [Array, String],
       default () {
         return () => {}
       }
@@ -21,5 +23,21 @@ export default {
       type: Boolean,
       default: false
     }
+  },
+  computed: {
+    ...mapGetters(['courseContentIdRouteMap']),
+    onFinish () {
+      if (Array.isArray(this.nextContent)) {
+        return () => {
+          const routes = []
+          this.nextContent.forEach((id: string) => {
+            routes.push(() => this.$router.push({ path: this.courseContentIdRouteMap[id] }))
+          })
+        }
+      } else {
+        return () => this.$router.push({ path: this.courseContentIdRouteMap[this.nextContent] })
+      }
+    }
+
   }
 }
